Move LATMPCN so scons compiles
