+++ selfdrive/controls/lib/lateral_mpc_lib/lat_mpc.py
@@ -5,7 +5,6 @@ import numpy as np
 from casadi import SX, vertcat, sin, cos
 
 from common.realtime import sec_since_boot
-from selfdrive.controls.lib.drive_helpers import LAT_MPC_N as N
 from selfdrive.modeld.constants import T_IDXS
 
 if __name__ == '__main__':  # generating code
@@ -18,6 +17,7 @@ EXPORT_DIR = os.path.join(LAT_MPC_DIR, "c_generated_code")
 JSON_FILE = os.path.join(LAT_MPC_DIR, "acados_ocp_lat.json")
 X_DIM = 4
 P_DIM = 2
+N = 16
 COST_E_DIM = 3
 COST_DIM = COST_E_DIM + 1
 MODEL_NAME = 'lat'
